[Edit]AddFireballDelay
불 시전 시 딜레이 시간 추가, 플레이어 공격 함수 따로 추가

+++ Assets/Scripts/Player/FireBall.cs
@@ -14,7 +14,7 @@ public class FireBall : MonoBehaviour
     {
         if (Time.time > nextShotTime)
         {
-            nextShotTime = Time.time + msBetweenShots / 1000;
+            nextShotTime = (int)Time.time + msBetweenShots / 1000;
             Projectile newProjectile = Instantiate(projectile, muzzle.position, muzzle.rotation) as Projectile;
             newProjectile.SetSpeed(muzzleVelocity);
         }
